+++ project/src/components/review-form/review-form.tsx
@@ -2,7 +2,7 @@ import { FormEvent, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate, useParams } from 'react-router';
 import { RootState } from '../../store/reducer';
-import { publishReview } from '../../store/slice/reviewsReducer';
+import { publishReview } from '../../store/actions/reviewsActions';
 import style from './review-form.module.scss';
 
 const STARS = [...Array(11).keys()].slice(1);
